[RENAME] domain 패키지 명 user로 변경 #4
