--- src/main/java/org/sopkathon/www/DoSoptSopkathon/domain/domain/entity/Gender.java
@@ -1,5 +1,0 @@
-package org.sopkathon.www.DoSoptSopkathon.domain.domain.entity;
-
-public enum Gender {
-   FEMALE, MALE
-}
