+++ src/main/java/org/sopkathon/www/DoSoptSopkathon/user/domain/entity/Gender.java
@@ -1,0 +1,5 @@
+package org.sopkathon.www.DoSoptSopkathon.user.domain.entity;
+
+public enum Gender {
+   FEMALE, MALE
+}
