+++ src/main/java/org/sopkathon/www/DoSoptSopkathon/user/domain/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package org.sopkathon.www.DoSoptSopkathon.user.domain.repository;
+
+import org.sopkathon.www.DoSoptSopkathon.user.domain.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
