+++ src/main/java/org/sopkathon/www/DoSoptSopkathon/user/service/UserService.java
@@ -1,9 +1,9 @@
-package org.sopkathon.www.DoSoptSopkathon.domain.service;
+package org.sopkathon.www.DoSoptSopkathon.user.service;
 
 import lombok.RequiredArgsConstructor;
-import org.sopkathon.www.DoSoptSopkathon.domain.controller.dto.request.CreateUserRequest;
-import org.sopkathon.www.DoSoptSopkathon.domain.domain.entity.User;
-import org.sopkathon.www.DoSoptSopkathon.domain.domain.repository.UserRepository;
+import org.sopkathon.www.DoSoptSopkathon.user.controller.dto.request.CreateUserRequest;
+import org.sopkathon.www.DoSoptSopkathon.user.domain.entity.User;
+import org.sopkathon.www.DoSoptSopkathon.user.domain.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
